+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 * Support DEC synchronized updates with Alacritty-compatible timeout recovery.
+* Report terminal view focus changes for DEC focus tracking mode.
 
 ## [4.0.0] - 2024-02-27
 * Update for Flutter 3.19 [#190]. Thanks [@domesticmouse].